requirements.txt
pywebview
winsdk
flask
flask-socketio
keyboard
Pillow